git-svn-id: http://qon.jwdt.org/var/svn/catracker@202 97746f1a-7bc6-4ed5-8832-28bd89886fd8

+++ GPU/TPCCAGPUTracking/tracking-ca/standalone/code/AliHLTTPCCASliceData.cxx
@@ -395,8 +395,8 @@ void AliHLTTPCCASliceData::ClearHitWeights()
 
 #ifdef ENABLE_VECTORIZATION
   const int_v v0( Zero );
-  const float *const end = fHitWeights + fNumberOfHits;
-  for ( float *mem = fHitWeights; mem < end; mem += v0.Size ) {
+  const int *const end = fHitWeights + fNumberOfHits;
+  for ( int *mem = fHitWeights; mem < end; mem += v0.Size ) {
     v0.store( mem );
   }
 #else
